terrain_ompl_rrt: move GridMap and TerrainMap to own module
